+++ clients/kafka-common/src/main/java/org/eclipse/hono/client/kafka/producer/KafkaProducerFactory.java
@@ -13,6 +13,9 @@
 
 package org.eclipse.hono.client.kafka.producer;
 
+import java.time.Duration;
+
+import org.apache.kafka.clients.CommonClientConfigs;
 import org.eclipse.hono.client.kafka.metrics.KafkaClientMetricsSupport;
 
 import io.vertx.core.Future;
@@ -46,8 +49,36 @@ public interface KafkaProducerFactory<K, V> {
      * @param producerName The name to identify the producer.
      * @param config The Kafka configuration with which the producer is to be created.
      * @return an existing or new producer.
+     * @throws org.apache.kafka.common.KafkaException If creating the producer failed because of an invalid config or
+     *             because bootstrap servers could not be resolved.
      */
     KafkaProducer<K, V> getOrCreateProducer(String producerName, KafkaProducerConfigProperties config);
+
+    /**
+     * Gets a producer for sending data to Kafka.
+     * <p>
+     * If producer creation fails because the {@value CommonClientConfigs#BOOTSTRAP_SERVERS_CONFIG} config property
+     * contains a (non-empty) list of URLs that are not (yet) resolvable, further creation attempts are done with some
+     * delay in between. These retries are done until the given <em>retriesTimeout</em> has elapsed.
+     * <p>
+     * The producer returned may be either newly created or it may be an existing producer for the given producer name.
+     * The config parameter might be ignored if an existing producer is returned.
+     * <p>
+     * Do not hold references to the returned producer between send operations, because the producer might be closed by
+     * the factory. Instead, always get an instance by invoking this method.
+     * <p>
+     * When the producer is no longer needed, it should be closed to release the resources (like connections). NB: Do
+     * not close the returned producer directly. Instead, call {@link #closeProducer(String)} to let the factory close
+     * the producer and update its internal state.
+     *
+     * @param producerName The name to identify the producer.
+     * @param config The Kafka configuration with which the producer is to be created.
+     * @param retriesTimeout The maximum time for which retries are done. Using a negative duration or {@code null}
+     *                       here is interpreted as an unlimited timeout value.
+     * @return A future with an existing or new producer.
+     */
+    Future<KafkaProducer<K, V>> getOrCreateProducerWithRetries(String producerName,
+            KafkaProducerConfigProperties config, Duration retriesTimeout);
 
     /**
      * Closes the producer with the given producer name if it exists.