+++ clients/kafka-common/src/test/java/org/eclipse/hono/client/kafka/producer/KafkaProducerConfigPropertiesTest.java
@@ -19,6 +19,7 @@ import static com.google.common.truth.Truth.assertThat;
 
 import java.util.Map;
 
+import org.apache.kafka.clients.CommonClientConfigs;
 import org.apache.kafka.common.serialization.StringSerializer;
 import org.eclipse.hono.client.kafka.AbstractKafkaConfigProperties;
 import org.eclipse.hono.client.kafka.CommonKafkaClientConfigProperties;
@@ -84,7 +85,7 @@ public class KafkaProducerConfigPropertiesTest {
 
         assertThat(config.isConfigured()).isFalse();
 
-        config.setProducerConfig(Map.of(AbstractKafkaConfigProperties.PROPERTY_BOOTSTRAP_SERVERS, "kafka"));
+        config.setProducerConfig(Map.of(CommonClientConfigs.BOOTSTRAP_SERVERS_CONFIG, "kafka"));
         assertThat(config.isConfigured()).isTrue();
     }